Provide a base class containing shared code for action plugins that change membership state.
