+++ src/Plugin/Action/BlockOgMembership.php
@@ -2,14 +2,7 @@
 
 namespace Drupal\og\Plugin\Action;
 
-use Drupal\Core\Access\AccessResult;
-use Drupal\Core\Action\ActionBase;
-use Drupal\Core\Plugin\ContainerFactoryPluginInterface;
-use Drupal\Core\Session\AccountInterface;
-use Drupal\og\Entity\OgMembership;
-use Drupal\og\OgAccessInterface;
 use Drupal\og\OgMembershipInterface;
-use Symfony\Component\DependencyInjection\ContainerInterface;
 
 /**
  * Blocks a group membership.
@@ -20,69 +13,20 @@ use Symfony\Component\DependencyInjection\ContainerInterface;
  *   type = "og_membership"
  * )
  */
-class BlockOgMembership extends ActionBase implements ContainerFactoryPluginInterface {
+class BlockOgMembership extends ChangeOgMembershipStateBase {
 
   /**
-   * The OG access service.
-   *
-   * @var \Drupal\og\OgAccessInterface
+   * {@inheritdoc}
    */
-  protected $ogAccess;
-
-  /**
-   * Constructs an BlockOgMembership object.
-   *
-   * @param array $configuration
-   *   A configuration array containing information about the plugin instance.
-   * @param string $plugin_id
-   *   The plugin_id for the plugin instance.
-   * @param mixed $plugin_definition
-   *   The plugin implementation definition.
-   * @param \Drupal\og\OgAccessInterface $og_access
-   *   The OG access service.
-   */
-  public function __construct(array $configuration, $plugin_id, $plugin_definition, OgAccessInterface $og_access) {
-    parent::__construct($configuration, $plugin_id, $plugin_definition);
-    $this->ogAccess = $og_access;
+  public function getTargetState() {
+    return OgMembershipInterface::STATE_BLOCKED;
   }
 
   /**
    * {@inheritdoc}
    */
-  public static function create(ContainerInterface $container, array $configuration, $plugin_id, $plugin_definition) {
-    return new static(
-      $configuration,
-      $plugin_id,
-      $plugin_definition,
-      $container->get('og.access')
-    );
-  }
-
-  /**
-   * {@inheritdoc}
-   */
-  public function execute(OgMembership $membership = NULL) {
-    if (!$membership) {
-      return;
-    }
-    $membership->setState(OgMembershipInterface::STATE_BLOCKED)->save();
-  }
-
-  /**
-   * {@inheritdoc}
-   */
-  public function access($object, AccountInterface $account = NULL, $return_as_object = FALSE) {
-    /** @var \Drupal\og\Entity\OgMembership $object */
-    // Deny access if the membership is already blocked.
-    if ($object->getState() === OgMembershipInterface::STATE_BLOCKED) {
-      $access = AccessResult::forbidden();
-    }
-    else {
-      // Only grant access if the user can manage members in this group.
-      $access = $this->ogAccess->userAccess($object->getGroup(), 'manage members', $account);
-    }
-
-    return $return_as_object ? $access : $access->isAllowed();
+  public function getOriginalState() {
+    return NULL;
   }
 
 }
